Add initial states to the game.
Add a 'preloader' state for starting asset loading, and a 'menu' state
to start the game off with.

To make sure the sequence is working correctly, we load a test asset in
the 'proloader' state, and display it in the 'menu' state, so we know
the sequence does what it's expected to and things and flowing
correctly.